@@ -4,19 +4,12 @@
   "private": true,
   "dependencies": {
     "@heroicons/react": "^1.0.6",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^28.1.4",
-    "@types/node": "^18.0.3",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "bootstrap": "^4.5.2",
+    "bootstrap": "^4.6.1",
     "firebase": "^8.10.1",
-    "react": "^16.13.1",
-    "react-bootstrap": "^1.3.0",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-bootstrap": "^1.6.5",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^5.3.3",
     "react-scripts": "^5.0.1",
     "typescript": "^4.7.4"
   },
@@ -42,6 +35,13 @@
     ]
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.5.0",
+    "@testing-library/user-event": "^7.2.1",
+    "@types/jest": "^28.1.4",
+    "@types/node": "^18.0.3",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/react-router-dom": "^5.3.3"
   }
 }
