@@ -1,12 +1,27 @@
-import React from "react";
-import ReactDOM from "react-dom";
+// import React from "react";
+// import ReactDOM from "react-dom";
+// import "bootstrap/dist/css/bootstrap.min.css";
+
+// import App from "./components/App";
+
+// ReactDOM.render(
+//   <React.StrictMode>
+//     <App />
+//   </React.StrictMode>,
+//   document.getElementById("root")
+// );
+
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 import "bootstrap/dist/css/bootstrap.min.css";
 
-import App from "./components/App";
+import App from './components/App';
 
-ReactDOM.render(
-  <React.StrictMode>
+const rootElement = document.getElementById('root');
+const root = createRoot(rootElement!); // createRoot(container!) if you use TypeScript
+
+root.render(
+  <StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </StrictMode>
 );
